styling on login, signup and contact page, debugged population of book images
