+++ client/public/images/brave-new-world.jpg
