+++ client/public/images/great-gatsby.jpg
