+++ client/src/components/Contact.jsx
@@ -3,7 +3,7 @@ import React from "react";
 const Contact = () => {
   return (
     <>
-      <section className="contact-form">
+      <section >
       </section>
     </>
   );
