+++ client/vite.config.js
@@ -13,11 +13,11 @@ export default defineConfig({
         changeOrigin: true,
         secure: false,
       },
-      '/apiKey': {
-        target: 'http://localhost:3001',
-        changeOrigin: true,
-        secure: false,
-      },
+      // '/apiKey': {
+      //   target: 'http://localhost:3001',
+      //   changeOrigin: true,
+      //   secure: false,
+      // },
     }
   }
 })
